feat : 인터페이스와 구현체 나누기
