+++ README.md
@@ -9,3 +9,5 @@
 - [x] 특수 문자 가능: ( ), [ ], +, -, &, /, _ 그 외 특수 문자 사용 불가
 - [x] "카카오"가 포함된 문구는 담당 MD와 협의한 경우에만 사용할 수 있다.
 2. 전체적인 리팩토링
+- [x] 디렉토리 재설정 및 코드 가독성 향상
+- [x] repository - 인터페이스와 구현체 나누기
